dev-requirements.txt
click==7.0
chardet==3.0.4
datadog==0.22.0
elasticsearch==5.2.0
hiredis==0.2.0
pyodbc==4.0.22
python-dateutil==2.6.1
PyYAML==3.13
redis==2.10.6
requests==2.19.1
urllib3==1.23
orjson==2.1.1
pydantic==1.0
python-dotenv==0.10.3
# tests
coverage==4.3.4
mock==2.0.0
pytest==3.0.6
pytest-cov==2.4.0
tox==2.6.0
boto3==1.10.28
# alerting
raven==6.5.0